developing : > making navigation drawer > changing action bar into tool bar ToDo : > inserting pictures > time, date manual setting > making other fragments
