+++ app/app.iml
@@ -148,7 +148,6 @@
     <orderEntry type="library" name="Gradle: com.android.support:support-v4-28.0.0" level="project" />
     <orderEntry type="library" name="Gradle: com.android.support:viewpager-28.0.0" level="project" />
     <orderEntry type="library" name="Gradle: com.android.support:support-annotations:28.0.0@jar" level="project" />
-    <orderEntry type="library" name="Gradle: com.google.android.gms:play-services-location-11.0.1" level="project" />
     <orderEntry type="library" name="Gradle: com.google.android.gms:play-services-tasks-15.0.1" level="project" />
     <orderEntry type="library" name="Gradle: com.google.api-client:google-api-client:1.22.0@jar" level="project" />
     <orderEntry type="library" name="Gradle: com.google.apis:google-api-services-calendar:v3-rev235-1.22.0@jar" level="project" />
