+++ app/build.gradle
@@ -34,8 +34,6 @@ dependencies {
         exclude group: 'org.apache.httpcomponents'
     }
 
-    implementation "com.google.android.gms:play-services-location:11.0.1"
-
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
