+++ app/src/main/res/menu/navigation_drawer_menu.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:tools="http://schemas.android.com/tools"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <group android:checkableBehavior="single">
+        <item android:title="다이어리">
+            <menu>
+                <item
+                    android:id="@+id/calendar"
+                    android:icon="@color/colorPrimary"
+                    android:title="달력" />
+                <item android:title="reserved" />
+            </menu>
+        </item>
+        <item android:title="메모">
+            <menu>
+                <item
+                    android:id="@+id/show"
+                    android:icon="@color/colorPrimary"
+                    android:title="보기"/>
+                <item
+                    android:id="@+id/search"
+                    android:icon="@color/colorPrimary"
+                    android:title="검색"/>
+            </menu>
+        </item>
+        <item
+            android:id="@+id/settings"
+            android:icon="@color/colorPrimary"
+            android:title="설정" />
+    </group>
+</menu>
